+++ CHANGELOG.md
@@ -1,7 +1,8 @@
-* Feature - adding geo properties to the mdtatdata config object
+* Feature - adding integrity check to Threshold Profile deletion
 ## Current Release 
 ### 0.285.0 
 **Release Date:** Thu Jan 24 14:22:06 UTC 2019     
+* Feature - adding geo properties to the metatdata config object
 ## Previous Releases 
 ### 0.284.0 
 **Release Date:** Mon Jan 21 17:37:48 UTC 2019